fix(tracing): move date window normalizer into presentation layer
The import-linter "presentation must use facade" contract forbids the
presentation layer from importing new top-level backend modules. The window
normalizer is DRF-aware request validation (it raises a DRF ValidationError),
so it belongs in the presentation package rather than behind the facade, which
must stay free of HTTP concerns. Move it there — presentation→presentation
imports are allowlisted — and revert the serializer help_text to its original
wording to avoid drift in the committed OpenAPI/MCP generated artifacts.

Generated-By: PostHog Code
Task-Id: 81588816-332c-4fa0-8c26-84f0f973eaab
